# backend/src/main/resources/contracts/autospec-backend-v1.openapi.yaml
openapi: 3.0.3
info:
  title: AutoSpec Backend API
  version: v1
  description: Spring Boot API contract for AutoSpec project orchestration, artifacts, exports, and observability.
servers:
  - url: /
security:
  - SessionToken: []
paths:
...
  /api/auth/login:
    post:
      summary: Create an authenticated persistent session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Session token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/logout:
    post:
      summary: Revoke the current persistent session token.
      parameters:
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '204':
          description: Session revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/projects:
    get:
      summary: List projects visible to the session user.
      parameters:
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Project history visible to the session user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a project from a natural-language requirement.
      parameters:
        - $ref: '#/components/parameters/SessionToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '200':
          description: Created project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/projects/{projectId}/code-skeleton:
    post:
      summary: Generate and persist a code skeleton export for a project.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Generated code skeleton export payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeGenerationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Specification or build verification gate rejected the bundle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/projects/{projectId}/trace-graph:
    get:
      summary: Fetch stable artifact components and explicit trace edges for one workflow run.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
        - name: workflowRunId
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Artifact trace graph.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactTraceGraphResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/projects/{projectId}/code-generation-jobs:
    get:
      summary: List durable code generation jobs.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Code generation job metadata list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CodeGenerationJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/projects/{projectId}/code-generation-jobs/{jobId}/cancel:
    post:
      summary: Cancel a running code generation job.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Cancelled code generation job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeGenerationJobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/projects/{projectId}/code-generation-jobs/{jobId}/retry:
    post:
      summary: Retry a failed code generation job.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Retried code skeleton export payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeGenerationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    SessionToken:
      type: apiKey
      in: cookie
      name: AUTOSPEC_SESSION
  parameters:
    SessionToken:
      name: AUTOSPEC_SESSION
      in: cookie
      required: true
      schema:
        type: string
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    PaginationLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    PaginationOffset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Missing or invalid session token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    TooManyRequests:
      description: Login attempt rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Session user lacks the required project role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Project-scoped resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Operation conflicts with current durable workflow state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      required: [userId, username, displayName]
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        displayName:
          type: string
    CreateProjectRequest:
      type: object
      required: [name, requirement]
      properties:
        name:
          type: string
          minLength: 1
        requirement:
          type: string
          minLength: 1
    CreateProjectResponse:
      type: object
      required: [projectId, status]
      properties:
        projectId:
          type: integer
          format: int64
        status:
          type: string
    ProjectResponse:
      type: object
      required: [projectId, name, originalRequirement, status, createdAt, updatedAt]
      properties:
        projectId:
          type: integer
          format: int64
        name:
          type: string
        originalRequirement:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CodeGenerationResponse:
      type: object
      required: [format, content, fileName, mediaType, encoding]
      properties:
        format:
          type: string
        content:
          type: string
        fileName:
          type: string
        mediaType:
          type: string
        encoding:
          type: string
    CodeGenerationJobResponse:
      type: object
      required: [id, status, manifest, createdAt]
      properties:
        id:
          type: integer
          format: int64
        retryOfJobId:
          type: integer
          format: int64
          nullable: true
        status:
          type: string
        manifest:
          type: string
        gateStatus:
          type: string
          nullable: true
          enum: [PASSED, BLOCKED]
        manifestHash:
          type: string
          nullable: true
        verificationJson:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        verifiedAt:
          type: string
          format: date-time
          nullable: true
...
    ArtifactTraceGraphResponse:
      type: object
      required: [components, edges]
      properties:
        workflowRunId:
          type: integer
          format: int64
          nullable: true
        components:
          type: array
          items:
            $ref: '#/components/schemas/ArtifactTraceComponent'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/ArtifactTraceEdge'
    ArtifactTraceComponent:
      type: object
      required: [artifactId, artifactType, componentKey, componentType, displayName, jsonPath, contentHash]
      properties:
        artifactId:
          type: integer
          format: int64
        artifactType:
          type: string
        componentKey:
          type: string
        componentType:
          type: string
        displayName:
          type: string
        jsonPath:
          type: string
        contentHash:
          type: string
    ArtifactTraceEdge:
      type: object
      required: [sourceArtifactId, fromComponentKey, toComponentKey, relationType]
      properties:
        sourceArtifactId:
          type: integer
          format: int64
        fromComponentKey:
          type: string
        toComponentKey:
          type: string
        relationType:
          type: string
...
    ApiErrorResponse:
      type: object
      required: [code, status, message, path, timestamp, fieldErrors, details]
      properties:
        code:
          type: string
        status:
          type: integer
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        details:
          type: object
          additionalProperties:
            type: string